KB-12-3 rename
