--- src/main/kotlin/lesson_12_3/main.kt
@@ -1,7 +1,0 @@
-package lesson_12_3
-
-fun main() {
-    val sixthDayLife = KeepWeather(1, 10, atmosphericPressure = 660)
-
-    sixthDayLife.printWeather()
-}
